split stats funcions on other file
