+++ src/routes/index.js
@@ -2,4 +2,5 @@
 var items = require('./routesItems.js');
 var tasks = require('./routesTasks.js');
 var status = require('./routesStatus.js');
-module.exports = [].concat(tasks, items, status);
+var stats = require('./routesStats.js');
+module.exports = [].concat(tasks, items, status, stats);
