+++ src/routes/routesTasks.js
@@ -80,47 +80,4 @@ module.exports = [{
     },
     handler: ControllerTasks.deleteTasks
   }
-}, {
-  method: 'GET',
-  path: '/tasks/{idtask}/activity/{from}/{to}',
-  config: {
-    description: 'Return activity in the task',
-    validate: {
-      params: {
-        idtask: Joi.string().required(),
-        from: Joi.string().required(),
-        to: Joi.string().required()
-      }
-    },
-    handler: ControllerTasks.listTasksActivity
-  }
-}, {
-  method: 'GET',
-  path: '/tasks/{idtask}/activity/{user}/{from}/{to}',
-  config: {
-    description: 'Return the user activity in the task',
-    validate: {
-      params: {
-        idtask: Joi.string().required(),
-        user: Joi.string().required(),
-        from: Joi.string().required(),
-        to: Joi.string().required()
-      }
-    },
-    handler: ControllerTasks.listTasksActivityByUser
-  }
-}, {
-  method: 'GET',
-  path: '/tasks/{idtask}/track_stats/{from}/{to}',
-  config: {
-    description: 'Return the tracking stats',
-    validate: {
-      params: {
-        idtask: Joi.string().required(),
-        from: Joi.string().required(),
-        to: Joi.string().required()
-      }
-    },
-    handler: ControllerTasks.trackStats
-  }
 }];
